requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
plaid-python==9.1.0
pandas==2.1.3
APScheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
streamlit==1.28.0
plotly==5.17.0
numpy==1.24.3